Create css styling for header and hero section
The html and css code for the header and hero are in place and have
base functionality. Encountered a bug in which the space between
the hero image and the hero text becomes uncomfortably large
when extending the size of the browser window beyond a certain point.

Attempted to use justify-content: center on the hero section however
this causes a mismatch between the left and right margins between
the header and hero. Will need to find a solution that allows
the margin to grow in response to the size of the window while also
preventing an overabundance of space between the hero image and text.